requirements.txt
cycler==0.10.0
Cython==0.25.2
decorator==4.0.10
graphviz==0.5.2
hmmlearn==0.2.0
joblib==0.10.3
matplotlib==1.5.3
networkx==1.11
numpy==1.11.2
Pillow==3.4.2
pomegranate==0.6.1
pydot-ng==1.0.0
pygraphviz==1.3.1
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
scikit-learn==0.18.1
scipy==0.18.1
seqlearn==0.2
six==1.10.0
python-weka-wrapper3==0.1.2